% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws-utils.R
\name{aws-utils}
\alias{aws-utils}
\alias{check_aws_cli}
\alias{check_aws_creds}
\title{AWS utility functions}
\usage{
check_aws_cli()

check_aws_creds()
}
\description{
\itemize{
\item \code{check_aws_cli()} - checks that the installation path for \code{awscli}
exists (as returned by \code{Sys.which("aws")})
\item \code{check_aws_creds()} - checks that the relevant enviroemnets variables
for connecting to AWS are set
}
}
