% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller_fun_args.R
\name{caller_fun_args}
\alias{caller_fun_args}
\title{Caller function argument values}
\usage{
caller_fun_args()
}
\description{
Get a list of all argument values for the caller function, including
default argument values and any additional arguments passed via \link[base]{dots}.
}
\details{
I made this function because \link[base]{match.call} doesn't return default
argument values and I couldn't find an alternative that did in any other packages.
}
\examples{
\dontrun{

h = function(a = 1, b = 2, c = 3, ...) caller_fun_args()
h(b = 22, d = 44, e = 55)
}
}
