% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_r_version.R
\name{check_r_version}
\alias{check_r_version}
\title{Check R version}
\usage{
check_r_version()
}
\description{
Check to see if a later R version is available for download.
\itemize{
\item Prints the current R version to console
\item Checks CRAN to see if a later version is available and, if so, prints a message
to the console
\item Only works in interactive mode
}
}
