% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-versions.R
\name{check_r_version}
\alias{check_r_version}
\alias{check_rstudio_version}
\title{Check R version}
\usage{
check_r_version(check_for_updates = TRUE)

check_rstudio_version(check_for_updates = TRUE)
}
\arguments{
\item{check_for_updates}{if \code{FALSE} then only the current version is printed}
}
\description{
Check to see if a later R version is available for download.
\itemize{
\item Prints the current R/RStudio version to console
\item Checks to see if a later version is available and, if so, prints a message to
the console
\item Only works in interactive mode
}
}
