% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col2hex}
\alias{col2hex}
\title{Convert colours to hexadecimal format}
\usage{
col2hex(x)
}
\arguments{
\item{x}{vector of colours value (character or integer)}
}
\description{
Convert a vector of colours to hexadecimal format.
}
\examples{
col2hex("blue")
col2hex(4)
col2hex("4")
col2hex(c(4,"blue"))
}
