% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion-matrix.R
\name{confusion}
\alias{confusion}
\alias{tpr}
\alias{tnr}
\alias{fnr}
\alias{fpr}
\alias{recall}
\alias{sensitivity}
\alias{specificity}
\alias{support}
\alias{prevalence}
\alias{accuracy}
\alias{precision}
\alias{lift}
\alias{f1}
\title{Confusion matrix metrics}
\usage{
tpr(tp, fp, fn, tn)

tnr(tp, fp, fn, tn)

fnr(tp, fp, fn, tn)

fpr(tp, fp, fn, tn)

recall(tp, fp, fn, tn)

sensitivity(tp, fp, fn, tn)

specificity(tp, fp, fn, tn)

support(tp, fp, fn, tn)

prevalence(tp, fp, fn, tn)

accuracy(tp, fp, fn, tn)

precision(tp, fp, fn, tn)

lift(tp, fp, fn, tn)

f1(tp, fp, fn, tn)
}
\arguments{
\item{tp}{n true positive (numeric)}

\item{fp}{n false positive (numeric)}

\item{fn}{n false negative (numeric)}

\item{tn}{n true negative (numeric)}
}
\description{
Helper functions for calculating confusion matrix metrics.
}
\details{
TPR = TP / (TP + FN) = Recall = Sensitivity

FNR = FN / (TP + FN)

TNR = TN / (TN + FP) = Specificity

FPR = FP / (TN + FP)

Support = TP + FN

Prevalence = (TP + FN) / (TP + FP + FN + TN)

Accuracy = (TP + TN) / (TP + FP + FN + TN)

Precision = TP / (TP + FP)

Lift = Precision / Prevalence
}
\examples{
tp = 87
fn = 13
tn = 76
fp = 32
tpr(tp, fp, fn, tn)
tnr(tp, fp, fn, tn)
fpr(tp, fp, fn, tn)
fnr(tp, fp, fn, tn)
sensitivity(tp, fp, fn, tn)
specificity(tp, fp, fn, tn)
precision(tp, fp, fn, tn)
recall(tp, fp, fn, tn)
prevalence(tp, fp, fn, tn)
accuracy(tp, fp, fn, tn)
support(tp, fp, fn, tn)
f1(tp, fp, fn, tn)
}
