% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-eda.R
\name{df-eda}
\alias{df-eda}
\alias{skim}
\alias{sketch}
\title{Explore a data frame}
\usage{
skim(data, verbose = TRUE)

sketch(data, x, y, group, min_unique_continuous = 10,
  max_unique_discrete = 50, max_distinct_group = 3, x_bins = NULL,
  x_binwidth = NULL, x_col = "dodgerblue", y_points = NULL,
  y_method = NULL, y_family = NULL, y_col = "grey30", k = 10)
}
\arguments{
\item{data}{a data frame}

\item{verbose}{if \code{TRUE} messages will be printed to the console}

\item{x}{unquoted expression giving the variable to be plotted (can be a combination of
one of more variables in data)}

\item{y}{TODO}

\item{group}{optional grouping variable}

\item{min_unique_continuous}{integer, if \code{x} is non-discrete and the number of
distinct values in \code{x} is less than \code{min_distinct_xnum} a error will be
thrown (see Details)}

\item{max_unique_discrete}{integer, if \code{x} is discrete and the number of distinct
values in \code{x} exceeds \code{max_distinct_xcat} a error will be thrown (see
Details)}

\item{max_distinct_group}{integer, if the number of distinct values in \code{group}
exceeds \code{max_distinct_group} a error will be thrown}

\item{x_bins}{passed to \link[ggplot2]{geom_histogram} (ignored if \code{x} is
discrete)}

\item{x_binwidth}{TODO}

\item{x_col}{TODO}

\item{y_points}{TODO}

\item{y_method}{TODO}

\item{y_family}{TODO}

\item{y_col}{TODO}

\item{k}{TODO}
}
\description{
\itemize{

\item \code{skim()} - returns a \link[dplyr]{tibble} with rows corresponding to the
variables in \code{data} and columns giving values for a range of summary statistics
(the result is returned silently). When \code{verbose = TRUE}, messages are printed
to the console to provide high-level insight and highlight possible issues with data
that might warrent further inspection prior to analysis.

\item \code{sketch()} - plots the distribution of a variable (or combination of
variables) in \code{data} - the relationship with a second variable (or combination
of variables) can be overlayed and faceting by a thrid variable (or combination
of variables) is also possible.

}
}
\details{
\link[dplyr]{type_sum} is used to determine the class of the \code{x}
variable:
\itemize{
\item if \code{x} is one of \code{"chr"}, \code{"fct"}, \code{"lgl"} or
\code{"ord"} then it is treated as discrete
\item if \code{x} is one of \code{"dbl"} or \code{"int"} then it is treated as
non-discrete
\item no other variable types are supported for \code{x}
}

\link[ggplot2]{geom_histogram} is used to plot non-discrete x variables and
\link[ggplot2]{geom_bar} is used to plot discrete x variables.
}
\examples{
\dontrun{

library(tidyverse)
theme_set(theme_oddments())

# iris -----

as_tibble(iris)
results <- skim(iris)
print(results)

iris \%>\% sketch(Species)
iris \%>\% sketch(Species, Sepal.Length)
iris \%>\% sketch(Species, Sepal.Length, y_method = "glm")
iris \%>\% sketch(Species, Sepal.Length, y_method = "none", y_points = TRUE)
iris \%>\% sketch(Species, Sepal.Length, y_method = "glm", y_points = TRUE)

iris \%>\% sketch(Sepal.Length)
iris \%>\% sketch(Sepal.Length, x_binwidth = 0.5)
iris \%>\% sketch(Sepal.Length, x_bins = 50)
iris \%>\% sketch(bin(Sepal.Length, breaks = seq(4, 8, 0.25)))
iris \%>\% sketch(bin_pretty(Sepal.Length))

iris \%>\% sketch(Sepal.Length, Sepal.Width)
iris \%>\% sketch(Sepal.Length, Sepal.Width, y_points = TRUE)
iris \%>\% sketch(Sepal.Length, Sepal.Width, y_points = TRUE,  y_method = "glm")
iris \%>\% sketch(Sepal.Length, Sepal.Width, y_points = TRUE,  y_method = "gam")
iris \%>\% sketch(Sepal.Length, Sepal.Width, y_points = TRUE,  y_method = "none")
iris \%>\% sketch(Sepal.Length, Sepal.Width, Species)
iris \%>\% sketch(Sepal.Length, Sepal.Width, Species, y_points = TRUE)
iris \%>\% sketch(Sepal.Length, Sepal.Width, Species, y_points = TRUE,  y_method = "glm")
iris \%>\% sketch(Sepal.Length, Sepal.Width, Species, y_points = TRUE,  y_method = "gam")
iris \%>\% sketch(Sepal.Length, Sepal.Width, Species, y_points = TRUE,  y_method = "none")


# mtcars -----

as_tibble(mtcars)
results <- skim(mtcars)
print(results)

mtcars \%>\% sketch(mpg)
mtcars \%>\% sketch(disp)
mtcars \%>\% sketch(mpg, disp)
mtcars \%>\% sketch(mpg, disp, y_method = "gam")
mtcars \%>\% sketch(mpg, vs)
mtcars \%>\% sketch(mpg, vs, y_method = "glm")
mtcars \%>\% sketch(mpg, vs, y_method = "gam")

mtcars \%>\% sketch(bin_quantile(mpg, 5))
mtcars \%>\% sketch(bin_quantile(mpg, 100))

mtcars \%>\% sketch(bin_pretty(mpg))
mtcars \%>\% sketch(bin_pretty(mpg), vs)
mtcars \%>\% sketch(bin_pretty(mpg), vs, y_method = "glm")


# attenu -----

as_tibble(attenu)
results <- skim(attenu)
print(results)

attenu \%>\% sketch(dist, accel)
attenu \%>\% sketch(log(dist), accel)
attenu \%>\% sketch(log(dist), log(accel))
attenu \%>\% sketch(log(dist), log(accel), y_method = "glm")
attenu \%>\% sketch(log(dist), log(accel), y_method = "gam")


# nycflights13::airports -----

nycflights13::airports
results <- skim(nycflights13::airports)
print(results)


# nycflights13::planes -----

nycflights13::planes
results <- skim(nycflights13::planes)
print(results)


# nycflights13::weather -----

nycflights13::weather
results <- skim(nycflights13::weather)
print(results)

nycflights13::weather \%>\%
    select_if(is.numeric) \%>\% 
    filter(complete.cases(.)) \%>\% 
    sample_frac(0.05) \%>\% 
    pairs()

nycflights13::weather \%>\% 
    sample_frac(0.1) \%>\% 
    sketch(month, temp, y_method = "gam", x_bins = 12, x_binwidth = 0.5)

nycflights13::weather \%>\% 
    sample_frac(0.1) \%>\% 
    sketch(as.factor(month), temp, y_method = "glm")


# nycflights13::flights -----

nycflights13::flights
results <- skim(nycflights13::flights)
print(results)


mtcars \%>\% pairs

mtcars \%>\% sketch(mpg)
mtcars \%>\% sketch(mpg, disp)
mtcars \%>\% sketch(mpg, disp, am)
mtcars \%>\% sketch(mpg, disp, vs)
mtcars \%>\% sketch(mpg, disp, cyl)

mtcars \%>\% sketch(as.character(gear), mpg)
mtcars \%>\% sketch(as.character(gear), qsec)
mtcars \%>\% sketch(as.character(gear), am)
mtcars \%>\% sketch(as.character(gear), vs)

mtcars \%>\% sketch(gear, vs)
mtcars \%>\% sketch(gear, am)
mtcars \%>\% sketch(as.integer(gear), am)

mtcars \%>\% sketch(mpg, disp)
mtcars \%>\% sketch(mpg, disp, y_method = "glm")
mtcars \%>\% sketch(mpg, disp, y_method = "gam")

mtcars \%>\% sketch(mpg, vs)
mtcars \%>\% sketch(mpg, vs, y_method = "gam")

mtcars \%>\% sketch(cyl)
mtcars \%>\% sketch(cyl / disp)
mtcars \%>\% sketch(cyl / disp, vs + am)
}
}
