% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-utils.R
\name{df-utils}
\alias{df-utils}
\alias{df_kable}
\title{Data frame utils}
\usage{
df_kable(x, scroll = FALSE, ...)
}
\arguments{
\item{x}{(data.frame)}

\item{scroll}{(boolean) if \code{TRUE} then a \link[kableExtra]{scroll_box} will be
used
\link[knitr]{kable}}

\item{...}{additional arguments to pass to the underlying function (see Details)}
}
\description{
\itemize{
\item \code{df_kable()} - wraps \link[knitr]{kable}, \link[kableExtra]{kable_styling}
and \link[kableExtra]{scroll_box}. Dots get passed to \link[knitr]{kable}.
\item \code{df_datatable()} - wraps \link[DT]{datatable}
\item \code{df_print()} - wraps \link[base]{print.data.frame} and uses
defaults \code{row.names=FALSE} and \code{right=FALSE}
}

See: https://rstudio.github.io/DT
}
\examples{
\dontrun{

mtcars \%>\%
  head() \%>\%
  df_kable() \%>\%
  view_html()
}
}
