% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-friendly.R
\name{do_call}
\alias{do_call}
\title{Pipe-friendly wrapper for \code{do.call}}
\usage{
do_call(args, what, ...)
}
\arguments{
\item{args}{a \emph{list} of arguments to the function call.  The
    \code{names} attribute of \code{args} gives the argument names.}

\item{what}{either a function or a non-empty character string naming the
    function to be called.}

\item{...}{additional arguments to pass to do.call}
}
\description{
See documentation for \link[base]{do.call}.
}
\examples{
x = list(1, 2, 3, 4, 5)

# base function 
do.call(c, x)

# base function with built-in pipe  
x |> (\(x) do.call(c, x))()

# wrapper function with built-in pipe  
x |> do_call(c)
}
