% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-utils.R
\name{ggplot-utils}
\alias{ggplot-utils}
\alias{theme_oddments}
\alias{theme_base}
\alias{theme_random}
\alias{theme_dmp}
\alias{ggplot2_colours}
\alias{ggplot2_colors}
\title{ggplot utility functions}
\usage{
theme_oddments(base_size = 11, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22)

theme_base(base_size = 12, base_family = "Helvetica",
  base_line_size = base_size/22, base_rect_size = base_size/22)

theme_random(pointsize = 8, family = "serif", linesize = 0.25,
  legend = FALSE)

theme_dmp(base_size = 8, base_family = "Helvetica",
  base_line_size = base_size/22, base_rect_size = base_size/22)

ggplot2_colours(ncols)

ggplot2_colors(ncols)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{pointsize}{TODO}

\item{family}{TODO}

\item{linesize}{TODO}

\item{legend}{TODO}

\item{ncols}{number of colours to generate}
}
\description{
\itemize{
\item \code{theme_oddments()} - bespoke ggplot theme
\item \code{theme_base()} - experimental theme to to make ggplots look like R base
plots
\item \code{ggplot2_colours()} - returns a character vector of (hexadecimal) colour
definitions
}
}
\examples{
\dontrun{
library(tidyverse)
ncols <- 8
df <- tibble(x = sample(letters[1:ncols], 100, replace = TRUE))
plot <- df \%>\%
  ggplot() +
  geom_bar(aes(x, fill = x), col = 1) +
  facet_grid(. ~ x, scales = "free_x") +
  labs(title = "title", subtitle = "subtitle")
plot
plot + theme_bw()
plot + theme_oddments()
plot + theme_base()
plot + theme_random()
plot + theme_dmp()
barplot(table(df$x), col = ggplot2_colours(ncols))
}
}
