% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ggplot2_colours}
\alias{ggplot2_colours}
\title{\pkg{ggplot2} default hue palette}
\usage{
ggplot2_colours(n)
}
\arguments{
\item{n}{number of colours to return}
}
\value{
a character vector of hexadecimal format colours.
}
\description{
\pkg{ggplot2} default hue palette
}
\examples{
ggplot2_colours(2)
ggplot2_colours(3)
ggplot2_colours(4)
scales::show_col(ggplot2_colours(2))
scales::show_col(ggplot2_colours(3))
scales::show_col(ggplot2_colours(4))
}
