% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_utils.R
\name{kable2}
\alias{kable2}
\title{TODO}
\usage{
kable2(x, scroll = FALSE, ...)
}
\arguments{
\item{x}{(data.frame)}

\item{scroll}{(boolean) if \code{TRUE} then a \link[kableExtra]{scroll_box} will be
used}

\item{...}{additional arguments to pass to \link[knitr]{kable}}
}
\description{
A combination of \link[knitr]{kable}, \link[kableExtra]{kable_styling} and
\link[kableExtra]{scroll_box}.
}
\examples{
\dontrun{

mtcars \%>\%
  head() \%>\%
  kable2() \%>\%
  view_html()
}

}
