% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{oddments}
\alias{oddments}
\alias{oddments-package}
\title{oddments: miscellaneous functions}
\description{
A collection of various functions that I've found
    useful at one time or another but which dont fit in any of my other
    repositories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dkidney/oddments}
  \item Report bugs at \url{https://github.com/dkidney/oddments/issues}
}

}
\author{
\strong{Maintainer}: Darren Kidney \email{darrenkidney@yahoo.co.uk}

}
