% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_functions.R
\name{pkg_functions}
\alias{pkg_functions}
\alias{pkg_status}
\title{Package utility functions}
\usage{
pkg_status(pkgs, lib = .libPaths()[1], repos = getOption("repos"),
  type = .Platform$pkgType, installed = NULL, available = NULL)
}
\arguments{
\item{pkgs}{\code{[character vector]} package names}

\item{lib}{\code{[character string]} the location of R library tree to search}

\item{repos}{\code{[character string]} the URL of the repository to use}

\item{type}{\code{[character string]} the package type}

\item{installed}{either \code{NULL} or an object returned by
\link[utils]{installed.packages}}

\item{available}{either \code{NULL} or an object returned by
\link[utils]{available.packages}}

\item{which_deps}{\code{[character vector]} names of dependency packages to include}

\item{recursive}{???}

\item{reverse}{???}
}
\description{
\describe{
\item{pkg_status}{
Returns a character vector giving the status of each package - each element is one
of: \code{"base"}; \code{"installed"}; \code{"available"} or \code{NA} (i.e.
unknown status)
}
\item{pkg_deps}{
???
}
\item{pkg_revdeps}{
???
}
\item{pkg_updates}{
???
}
\item{pkg_upgrade}{
???
}
}
}
\details{
???
}
\examples{
\dontrun{

library(oddments)

pkgs = c(
    "abcd",
    "base",
    "devtools",
    "furrr",
    "knitr",
    "rmarkdown",
    "stats",
    "tools"
)

pkgs \%>\% pkg_status()

pkgs \%>\% pkg_deps()

pkgs \%>\% pkg_revdeps()

pkgs \%>\% pkg_updates()

pkgs \%>\% pkg_upgrade()
}
}
