% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\alias{heading}
\alias{bullet}
\alias{item}
\alias{success}
\alias{concern}
\alias{itemize}
\title{Progress messages}
\usage{
heading(...)

bullet(x = NULL, ...)

item(x = NULL, ...)

success(x = NULL, ...)

concern(x = NULL, ...)

itemize(.f, ..., .message = NULL, .timer = TRUE)
}
\arguments{
\item{...}{arguments that get passed to \link[cli]{cat_bullet} (or
\link[cli]{cat_bullet} for \code{heading})}

\item{x}{\link{string} progress message}

\item{.f}{TODO}

\item{.message}{TODO}

\item{.timer}{TODO}
}
\description{
TODO
}
\examples{
\dontrun{

"message" \%>\% heading()
"message" \%>\% bullet()
"message" \%>\% item()
"message" \%>\% success()
"message" \%>\% concern()

item("item", col = "blue")
item("item", col = "black")
item("item", col = "blue")
item("item", col = "cyan")
item("item", col = "green")
item("item", col = "magenta")
item("item", col = "red")
item("item", col = "white")
item("item", col = "yellow")
item("item", col = "grey")
item("item", col = "silver")

# itemize -----

f1 <- function() {
    Sys.sleep(2)
    message("hello")
    message("how are you")
    print(10)
    return(100)
}
f2 <- function() {
    message("hi")
    Sys.sleep(0.001)
    warning("whoops")
    warning("not again...")
    print(10)
    return(100)
}
f3 <- function(timer = FALSE) {
    force(timer)
    itemize(f1, .message = "checking something", .timer = timer)()
    itemize(f2, .message = "checking something else", .timer = timer)()
}
f3()
f3(TRUE)
}
}
