% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-magrittr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{set_attr}
\alias{set_class}
\alias{set_names}
\alias{set_colnames}
\alias{set_rownames}
\alias{\%$\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:exposition]{\%$\%}}, \code{\link[magrittr:aliases]{set_attr}}, \code{\link[magrittr:aliases]{set_class}}, \code{\link[magrittr:aliases]{set_colnames}}, \code{\link[magrittr:aliases]{set_names}}, \code{\link[magrittr:aliases]{set_rownames}}}
}}

