% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_null}
\alias{replace_null}
\title{Replace null values}
\usage{
replace_null(x, replacement = NULL)
}
\arguments{
\item{x}{input value}

\item{replacement}{default replacement value if \code{x} is \code{NULL}}
}
\description{
Replace \code{NULL} values with a default value. Wrapper for
\code{rlang::\link[rlang]{\%||\%}}.
}
\examples{
x = NULL
y = 9
x \%>\% replace_null(10)
y \%>\% replace_null(10)
}
