% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd-utils.R
\name{rmd_utils}
\alias{rmd_utils}
\alias{rmd_html_template}
\alias{rmd_insert_pagebreak}
\title{Rmarkdown utility functions}
\usage{
rmd_html_template(...)

rmd_insert_pagebreak()
}
\description{
\itemize{
\item \code{rmd_html_template()} -
\item \code{rmd_insert_pagebreak()} - inserts a page break in an html, pdf or docx
rmarkdown document
}
}
\details{
\preformatted{
```{r, echo=FALSE, results="asis"}
rmd_insert_pagebreak()
```
}
}
\examples{
\dontrun{

# TODO
}
}
