% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_chr}
\alias{round_chr}
\title{Round digits and convert to character}
\usage{
round_chr(x, digits = 0)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{integer indicating the number of decimal places}
}
\description{
A pipe-friendly wrapper for a specific use case of \link[base]{sprintf}.
}
\examples{
sprintf("\%.3f", pi)
pi \%>\% round_chr(3)
}
