% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim.R
\name{skim}
\alias{skim}
\title{Summarise a data frame}
\usage{
skim(x, verbose = TRUE)
}
\arguments{
\item{x}{a data frame}

\item{verbose}{if \code{TRUE} messages will be printed to the console}
}
\description{
Returns a \link[dplyr]{tibble} with rows corresponding to the variables in
\code{x} and columns giving values for a range of summary statistics (result is
returned silently).

When \code{verbose = TRUE}, comments are printed to the console to highlight possible
issues with data that might warrent further inspection prior to analysis.
}
\examples{
\dontrun{

library(tidyverse)

as_tibble(iris)
results <- skim(iris)
print(results)

as_tibble(mtcars)
results <- skim(mtcars)
print(results)

as_tibble(esoph)
results <- skim(esoph)
print(results)

as_tibble(attenu)
results <- skim(attenu)
print(results)

nycflights13::airports
results <- skim(nycflights13::airports)
print(results)

nycflights13::planes
results <- skim(nycflights13::planes)
print(results)

nycflights13::weather
results <- skim(nycflights13::weather)
print(results)

nycflights13::flights
results <- skim(nycflights13::flights)
print(results)
}
}
