% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-utils.R
\name{str-utils}
\alias{str-utils}
\alias{collapse_str}
\alias{commas}
\alias{round_str}
\alias{integer_str}
\alias{difftime_str}
\title{String utils}
\usage{
collapse_str(..., .sep = " ", .use_quotes = NULL, .width = NULL,
  .indent = 0, .exdent = 0)

commas(..., .use_quotes = NULL, .width = getOption("width"),
  .indent = 0, .exdent = 0)

round_str(x, digits = getOption("digits"))

integer_str(x)

difftime_str(x, digits = 1, units = "auto")
}
\arguments{
\item{...}{one or more scalars and/or vectors of items to be coerced to strings}

\item{.sep}{TODO}

\item{.use_quotes}{indicates whether elements should be quoted (if \code{NULL} and if
the \code{...} elements are character then quotes are used by default)}

\item{.width}{passed to \link[stringr]{str_wrap} as the \code{width} argument}

\item{.indent}{passed to \link[stringr]{str_wrap} as the \code{indent} argument}

\item{.exdent}{passed to \link[stringr]{str_wrap} as the \code{exdent} argument}

\item{x}{a numeric vector}

\item{digits}{\link{integer} number of decimal places}

\item{units}{\link{string} time units (one of "auto", "secs", "mins", "hours", "days",
"weeks"))}
}
\description{
\itemize{
\item \code{collapse_str()} - collapses the input into a single string, character
inputs are 'quoted' by default
\item \code{commas()} - is a wrapper for \code{collapse_str()} using \code{.sep =
    ", "}
\item \code{round_str()} - converts numeric/integer vector input into a character
vector using a given number of decimal places (it wraps \link[base]{sprintf})
\item \code{integer_str()} - converts numeric/integer vector input to integer
format and returns a character vector without using scientific notation
\item \code{difftime_str()} - converts a unit length \link[base]{difftime} object
into a string that resembles the output of \link[base]{print.difftime}
}
}
\examples{
cat(collapse_str(0:100))
cat(collapse_str(0:100, .width = 20, .indent = 5, .exdent = 5))

cat(collapse_str(letters, LETTERS))
cat(collapse_str(letters, LETTERS, .width = 20, .indent = 5, .exdent = 5))
cat(commas(0:100))
cat(commas(letters, LETTERS))
1 \%>\% round(3)
1 \%>\% sprintf("\%.3f", .)
1 \%>\% round_str(3)

pi \%>\% round(3)
pi \%>\% sprintf("\%.3f", .)
pi \%>\% round_str(3)
1e10 \%>\% integer_str()
start <- Sys.time()
stop <- Sys.time()
dt <- difftime(stop, start)
dt \%>\% print()
dt \%>\% format(units = "mins") # doesn't change units
dt \%>\% as.character()
dt \%>\% difftime_str()
dt \%>\% difftime_str(units = "secs")
dt \%>\% difftime_str(units = "mins")
}
