% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_packages.R
\name{upgrade_packages}
\alias{upgrade_packages}
\title{Upgrade packages}
\usage{
upgrade_packages(pkgs = NULL, lib = .libPaths()[1],
  repos = getOption("repos"), type = .Platform$pkgType,
  dependencies = c("Depends", "Imports", "LinkingTo"), quiet = TRUE)
}
\arguments{
\item{pkgs}{TODO}

\item{lib}{TODO}

\item{repos}{TODO}

\item{type}{TODO}

\item{dependencies}{TODO}

\item{quiet}{TODO}
}
\description{
\itemize{
\item Installs any packages in \code{pkgs} that are not already installed
\item Updates any packages in \code{pkgs} that are installed but out-of-date.
}
}
\examples{
\dontrun{

upgrade_packages("devtools")
upgrade_packages("tidyverse")
upgrade_packages("tidymodels")
}
}
