% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_html}
\alias{view_html}
\title{View html}
\usage{
view_html(x)
}
\arguments{
\item{x}{object representing html code (e.g. a string)}
}
\description{
View html code in RStudio viewer pane (requires \pkg{rstudioapi} to be
installed to access \code{rstudioapi::\link[rstudioapi]{viewer}})
}
\examples{
\dontrun{

mtcars \%>\%
  head \%>\%
  knitr::kable(format = "html") \%>\%
  view_html
}
}
