% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{col2hex}
\alias{col2hex}
\title{Convert colours to hexadecimal format}
\usage{
col2hex(x)
}
\arguments{
\item{x}{character or integer vector of colour values}
}
\value{
a character vector of hexadecimal format colours.
}
\description{
Convert a vector of colours to hexadecimal format.
}
\examples{
col2hex("blue")
col2hex(4)
col2hex("4")
col2hex(c(4,"blue"))
}
