% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-eda.R
\name{df-eda}
\alias{df-eda}
\alias{skim}
\alias{sketch}
\title{Explore a data frame}
\usage{
skim(data, verbose = TRUE)

sketch(data, x, y, group, min_unique_continuous = 10,
  max_unique_discrete = 50, max_distinct_group = 3, x_bins = 30,
  y_method = NULL, y_family = NULL, y_points = NULL, k = 10)
}
\arguments{
\item{data}{a data frame}

\item{verbose}{if \code{TRUE} messages will be printed to the console}

\item{x}{variable for which the distribution will be plotted}

\item{group}{optional grouping variable}

\item{max_distinct_group}{integer, if the number of distinct values in \code{group}
exceeds \code{max_distinct_group} a error will be thrown}

\item{min_distinct_xnum}{integer, if \code{x} is non-discrete and the number of
distinct values in \code{x} is less than \code{min_distinct_xnum} a error will be
thrown (see Details)}

\item{max_distinct_xcat}{integer, if \code{x} is discrete and the number of distinct
values in \code{x} exceeds \code{max_distinct_xcat} a error will be thrown (see
Details)}

\item{bins}{passed to \link[ggplot2]{geom_histogram} (ignored if \code{x} is
discrete)}
}
\description{
\itemize{

\item \code{skim()} - returns a \link[dplyr]{tibble} with rows corresponding to the
variables in \code{data} and columns giving values for a range of summary statistics
(the result is returned silently). When \code{verbose = TRUE}, messages are printed
to the console to provide high-level insight and highlight possible issues with data
that might warrent further inspection prior to analysis.

\item \code{sketch()} - plots the distribution of a variable (or combination of
variables) in \code{data} - the relationship with a second variable (or combination
of variables) can be overlayed and faceting by a thrid variable (or combination
of variables) is also possible.

}
}
\details{
\link[dplyr]{type_sum} is used to determine the class of the \code{x}
variable:
\itemize{
\item if \code{x} is one of \code{"chr"}, \code{"fct"}, \code{"lgl"} or
\code{"ord"} then it is treated as discrete
\item if \code{x} is one of \code{"dbl"} or \code{"int"} then it is treated as
non-discrete
\item no other variable types are supported for \code{x}
}

\link[ggplot2]{geom_histogram} is used to plot non-discrete x variables and
\link[ggplot2]{geom_bar} is used to plot discrete x variables.
}
