% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dttm-utils.R
\name{dttm-utils}
\alias{dttm-utils}
\alias{tz_get}
\alias{tz_set}
\alias{tz_conv}
\title{Date-time utility functions}
\usage{
tz_get(x)

tz_set(x, tz)

tz_conv(x, tz)
}
\arguments{
\item{x}{[\link[base]{POSIXct} or \link[base]{POSIXlt}] object}

\item{tz}{[string] the desired time zone (must be one of
\link[base]{OlsonNames})}
}
\value{
Returns a
}
\description{
\itemize{
\item \strong{\code{tz_get}} - extracts the timezone
\item \strong{\code{tz_set}} - changes the timezone
\item \strong{\code{tz_conv}} - changes the timezone and the clock time
}
These function are just wrappers for \link[lubridate]{tz}, \link[lubridate]{force_tz}
and \link[lubridate]{with_tz} but I find their names more intuitive and easier to
remember than the \pkg{lubridate} funtions.
}
\examples{
\dontrun{

(x = Sys.time())
tz_get(x)
tz_set(x, "Asia/Kathmandu")
tz_conv(x, "Asia/Kathmandu")
}
}
