% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-utils.R
\name{function-utils}
\alias{function-utils}
\alias{capture_args}
\title{Function utils}
\usage{
capture_args()
}
\description{
\itemize{
\item \code{capture_args()} - returns a list of all argument values supplied to the
caller function, including the default argument values that were not specified in the
call plus and any additional arguments passed via \link[base]{dots}.
(\link[base]{match.call} ignores default arguments)
}
}
