% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddments-package.R
\docType{package}
\name{oddments-package}
\alias{oddments}
\alias{oddments-package}
\title{oddments: a mixed bag of useful functions}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dkidney/oddments}
  \item Report bugs at \url{https://github.com/dkidney/oddments/issues}
}

}
\author{
\strong{Maintainer}: Darren Kidney \email{darrenkidney@yahoo.co.uk}

}
