% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{oddments-package}
\alias{oddments}
\alias{oddments-package}
\title{oddments: a mixed bag of functions}
\description{
The oddments package contains various functions
    that I've found useful at one time or another but which dont really
    fit in any of my other repositories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dkidney/oddments}
  \item Report bugs at \url{https://github.com/dkidney/oddments/issues}
}

}
\author{
\strong{Maintainer}: Darren Kidney \email{darrenkidney@yahoo.co.uk}

}
