% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\alias{heading}
\alias{bullet}
\alias{item}
\alias{success}
\alias{concern}
\alias{panic}
\alias{itemize}
\title{Progress messages}
\usage{
heading(...)

bullet(...)

item(...)

success(...)

concern(...)

panic(...)

itemize(.f, ..., .message = NULL, .timer = TRUE)
}
\arguments{
\item{...}{arguments that get passed to \link[cli]{cat_bullet} (or
\link[cli]{cat_bullet} for \code{heading})}

\item{.f}{TODO}

\item{.message}{TODO}

\item{.timer}{TODO}

\item{x}{\link{string} progress message}
}
\description{
TODO
}
\examples{
\dontrun{

heading("a", " message")
bullet("a", " message")
item("a", " message")
success("a", " message")
concern("a", " message")
panic("a", " message")

# itemize -----

f1 <- function() {
    Sys.sleep(2)
    message("hello")
    message("how are you")
    print(10)
    return(100)
}
f2 <- function() {
    message("hi")
    Sys.sleep(0.001)
    warning("whoops")
    warning("not again...")
    stop("this is bad!")
    print(10)
    return(100)
}
f3 <- function(timer = FALSE) {
    force(timer)
    itemize(f1, .message = "checking something", .timer = timer)()
    itemize(f2, .message = "checking something else", .timer = timer)()
}
f3()
f3(TRUE)
}
}
